Move dataset header selection to its own component
